moved randomgradient function to utils
